Refactor Farsi researcher
