+++ packages/yoastseo/src/languageProcessing/languages/fa/helpers/createBasicWordForms.js
