--- packages/yoastseo/src/languageProcessing/languages/fa/researches/keyphraseDistribution.js
@@ -1,10 +1,0 @@
-import { keyphraseDistributionResearcher } from "../../../researches/base/keyphraseDistribution";
-
-/**
- * Determines which portions of the text did not receive a lot of content words from keyphrase and synonyms.
- *
- * @inheritDoc keyphraseDistributionResearcher
- */
-export default function( paper, researcher ) {
-	return keyphraseDistributionResearcher( paper, researcher, true );
-}
